Added a CCetractor
Now extraction is much more robust
More extraction types are supported

+++ src/main/java/org/chaiware/recursive_archive_extractor/RecursiveArchiveExtractor.java
@@ -1,29 +1,22 @@
 package org.chaiware.recursive_archive_extractor;
 
-import org.rauschig.jarchivelib.Archiver;
-import org.rauschig.jarchivelib.ArchiverFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.CopyOption;
 import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
-import java.util.HashMap;
-import java.util.Map;
 
 /**
  * Goes over the source folder recursively and copies all files to the target preserving the original folder structure <br/>
  * In the case of an archive, it extracts its contents to the target<br/>
  * Archives within archives will also be extracted
- * */
+ */
 public class RecursiveArchiveExtractor {
     private static final Logger log = LoggerFactory.getLogger(RecursiveArchiveExtractor.class);
     static String archiveSources = "";
     static String extractedTarget = "";
-
-    static Map<String, Integer> extensionCount = new HashMap<>();
 
     public static void main(String[] args) throws Exception {
         usage(args);
@@ -49,7 +42,7 @@ public class RecursiveArchiveExtractor {
      */
     private static void handleFolder(File file) throws IOException {
         log.info("In folder: {}, Files: {}", file.getPath(), file.listFiles().length);
-        String destination = extractedTarget + file.getPath().replace(archiveSources, "");
+        File destination = new File(extractedTarget + file.getPath().replace(archiveSources, ""));
         initializeFolder(destination);
 
         File[] filesList = file.listFiles();
@@ -57,7 +50,6 @@ public class RecursiveArchiveExtractor {
             if (currentFile.isFile()) {
                 handleFile(currentFile, destination);
             } else { // Folder
-                reportAndInitExtensionCount();
                 handleFolder(currentFile);
             }
         }
@@ -68,59 +60,33 @@ public class RecursiveArchiveExtractor {
      * If it is a regular file then copy it to the target <br/>
      * If it is an archived file then extract it to the target
      */
-    private static void handleFile(File file, String destinationPath) throws IOException {
-        String fileExtension = file.getName().substring(file.getName().lastIndexOf(".") + 1);
-        extensionCount.put(fileExtension, extensionCount.getOrDefault(fileExtension, 0) + 1);
-
-        try {
-            Archiver archiver = ArchiverFactory.createArchiver(file); // Throws an exception when not an archive
-            String targetFolder = destinationPath + "\\" + file.getName().substring(0, file.getName().lastIndexOf("."));
-            initializeFolder(targetFolder);
-            File destinationFolder = new File(targetFolder);
-            archiver.extract(file, destinationFolder);
-            log.info("Extracted: {} files to {}", destinationFolder.listFiles().length, destinationFolder.getPath());
+    private static void handleFile(File file, File destinationPath) throws IOException {
+        if (CCExtractor.isArchive(file.toPath())) {
+            File destinationArchiveFolder = new File(destinationPath + "\\" + file.getName().substring(0, file.getName().lastIndexOf(".")).replace("\\\\", "\\"));
+            initializeFolder(destinationArchiveFolder);
+            CCExtractor.extract(file.toPath(), destinationArchiveFolder);
+            log.info("Extracted: {} files to {}", destinationArchiveFolder.listFiles().length, destinationArchiveFolder.getPath());
 
             // Going over extracted files in order to find archives in them
-            File[] filesList = destinationFolder.listFiles();
-            for (File currentFile : filesList) {
-                if (currentFile.isFile()) {
-                    try {
-                        ArchiverFactory.createArchiver(currentFile); // Throws an exception when not an archive
-                        handleFile(currentFile, destinationFolder.getPath());
-                        currentFile.delete();
-                    } catch (Exception ex) { // This is ok, just continue
-
-                    }
+            for (File currentFile : destinationArchiveFolder.listFiles()) {
+                if (currentFile.isFile() && CCExtractor.isArchive(currentFile.toPath())) {
+                    handleFile(currentFile, destinationArchiveFolder);
+                    currentFile.delete();
                 }
             }
-        } catch (Exception ex) { // Not a compressed file
+        } else { // is a regular file
             Files.copy(file.toPath(), new File(destinationPath + "\\" + file.getName()).toPath(), StandardCopyOption.REPLACE_EXISTING);
         }
     }
 
-    private static File initializeFolder(String folderName) {
-        folderName = folderName.replace("\\\\", "\\");
-        File targetFolder = new File(folderName);
+    private static File initializeFolder(File targetFolder) {
         if (!targetFolder.exists()) {
             if (targetFolder.mkdirs()) {
-                log.info("Folder {} Created", targetFolder);
+                log.info("Folder {} Created", targetFolder.getPath());
             }
         }
 
         return targetFolder;
-    }
-
-    /**
-     * Logs a report of the current folder extension count then initializes the extension count
-     */
-    private static void reportAndInitExtensionCount() {
-        log.info("Extension Count Report:");
-        for (Map.Entry<String, Integer> entry: extensionCount.entrySet()) {
-            log.info("{}: {}", entry.getKey(), entry.getValue() );
-        }
-        log.info("\n");
-
-        extensionCount = new HashMap<>();
     }
 
     private static File initializeSourceFolder() throws Exception {
